requirements.txt
asgiref==3.7.2
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
crispy-bootstrap4==2022.1
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.57.0
django-axes==6.0.5
django-countries==7.2.1
django-crispy-forms==2.0
django-otp==1.2.2
django-ratelimit==4.0.0
docutils==0.20.1
Flask==3.0.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Kivy==2.2.1
Kivy-Garden==0.1.5
MarkupSafe==2.1.3
oauthlib==3.2.2
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pygame==2.5.2
Pygments==2.15.1
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stripe==6.7.0
typing_extensions==4.8.0
urllib3==1.26.16
Werkzeug==3.0.1
